pyproject.toml
[project]
name = "agent-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.10.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "python-dotenv>=1.1.1",
    "langfuse>=2.60.0",
    "pinecone-client>=4.1.0",
    "langchain-pinecone>=0.2.0",
    "langchain-community>=0.3.0",
    "numexpr>=2.10.1",
    "langchain>=1.2.12",
]
